+++ model_navigator/framework_api/commands/convert/onnx.py
@@ -26,7 +26,7 @@ from model_navigator.framework_api.logger import LOGGER
 from model_navigator.framework_api.runners.onnx import OnnxrtRunner
 from model_navigator.framework_api.utils import Framework, Status, format_to_relative_model_path, get_package_path
 from model_navigator.model import Format
-from model_navigator.utils import device, tensorrt
+from model_navigator.utils import devices, tensorrt
 
 
 class ConvertONNX2TRT(ConvertBase):
@@ -62,7 +62,7 @@ class ConvertONNX2TRT(ConvertBase):
         **kwargs,
     ) -> Optional[Path]:
         LOGGER.info("ONNX to TRT conversion started")
-        if not device.get_available_gpus():
+        if not devices.get_available_gpus():
             raise RuntimeError("No GPUs available.")
 
         if framework == Framework.PYT: