+++ model_navigator/framework_api/commands/convert/pyt.py
@@ -28,7 +28,7 @@ from model_navigator.framework_api.exceptions import ExecutionContext
 from model_navigator.framework_api.logger import LOGGER
 from model_navigator.framework_api.utils import JitType, Status, get_package_path, parse_kwargs_to_cmd
 from model_navigator.model import Format
-from model_navigator.utils.device import get_available_gpus
+from model_navigator.utils import devices
 
 
 class ConvertTorchScript2TorchTensorRT(ConvertBase):
@@ -56,7 +56,7 @@ class ConvertTorchScript2TorchTensorRT(ConvertBase):
         **kwargs,
     ) -> Optional[Path]:
         LOGGER.info("Conversion TorchScript to TorchTensorRT started")
-        if not get_available_gpus():
+        if not devices.get_available_gpus():
             raise RuntimeError("No GPUs available.")
 
         import torch_tensorrt  # pytype: disable=import-error # noqa: F401