+++ model_navigator/framework_api/contrib/huggingface/tensorflow/export.py
@@ -51,9 +51,9 @@ from model_navigator.framework_api.utils import (
     format2runtimes,
     get_default_max_workspace_size,
     get_default_workdir,
-    parse_enum,
 )
 from model_navigator.model import Format
+from model_navigator.utils import enums
 
 
 class HFDataLoader:
@@ -183,9 +183,9 @@ def export(
         profiler_config = ProfilerConfig()
 
     target_formats, target_precisions, runtimes = (
-        parse_enum(target_formats, Format),
-        parse_enum(target_precisions, TensorRTPrecision),
-        parse_enum(runtimes, RuntimeProvider),
+        enums.parse(target_formats, Format),
+        enums.parse(target_precisions, TensorRTPrecision),
+        enums.parse(runtimes, RuntimeProvider),
     )
     config = Config(
         Framework.TF2,