+++ model_navigator/framework_api/contrib/huggingface/torch/export.py
@@ -56,9 +56,9 @@ from model_navigator.framework_api.utils import (
     format2runtimes,
     get_default_max_workspace_size,
     get_default_workdir,
-    parse_enum,
 )
 from model_navigator.model import Format
+from model_navigator.utils import enums
 
 # pytype: enable=import-error
 
@@ -210,11 +210,11 @@ def export(
         profiler_config = ProfilerConfig()
 
     target_formats, jit_options, target_precisions, runtimes, precision_mode = (
-        parse_enum(target_formats, Format),
-        parse_enum(jit_options, JitType),
-        parse_enum(target_precisions, TensorRTPrecision),
-        parse_enum(runtimes, RuntimeProvider),
-        *parse_enum(precision_mode, TensorRTPrecisionMode),
+        enums.parse(target_formats, Format),
+        enums.parse(jit_options, JitType),
+        enums.parse(target_precisions, TensorRTPrecision),
+        enums.parse(runtimes, RuntimeProvider),
+        *enums.parse(precision_mode, TensorRTPrecisionMode),
     )
 
     config = Config(