+++ tests/unit_framework/test_dataloaders_tf2.py
@@ -18,7 +18,7 @@ from pathlib import Path
 import tensorflow
 
 import model_navigator as nav
-from model_navigator.utils.device import get_gpus
+from model_navigator.utils.devices import get_gpus
 
 # pytype: enable=import-error